requirements.txt
kafka-python==2.2.0
numpy==2.2.5
scipy==1.15.2
